feat(mnonorepo): project paths
