+++ packages/market-matcher/package.json
@@ -12,11 +12,11 @@
         "url": "https://github.com/energywebfoundation/ewf-monorepo/issues"
     },
     "version": "1.1.0",
-    "main": "dist/js/src/exports.js",
+    "main": "dist/js/src/index.js",
     "scripts": {
         "build": "scripts/build.js",
-        "build-ts": "rm -rf dist && tsc",
-        "debug": "node --inspect-brk dist/js/src/index.js",
+        "build-ts": "rm -rf dist && tsc -b tsconfig.build.json",
+        "debug": "node --inspect-brk dist/js/src/main.js",
         "generate-conf-schema": "typescript-json-schema --required --ignoreErrors tsconfig.json IMatcherConf > schemas/conf.schema.json",
         "generate-rule-schema": "typescript-json-schema --required --ignoreErrors tsconfig.json IRuleConf > schemas/rule.schema.json",
         "generate-schemas": "rm -rf schemas/ && mkdir schemas/ && npm run generate-sim-flow-schema && npm run generate-rule-schema && npm run generate-conf-schema",
@@ -24,12 +24,13 @@
         "lint": "tslint 'src/**/*{.ts,.tsx}'",
         "lint-fix": "tslint --fix 'src/**/*{.ts,.tsx}'",
         "prepare": "scripts/build.js",
-        "start": "npm run build && node dist/js/src/index.js",
+        "start": "npm run build && node dist/js/src/main.js",
         "start-ganache": "ganache-cli -m 'chalk park staff buzz chair purchase wise oak receive avoid avoid home' -l 8000000 -e 1000000 -a 20",
         "start-test-backend": "node ../../node_modules/@energyweb/utils-testbackend/dist/js/src/index.js",
         "prettier": "prettier --write --config-precedence file-override './src/**/*'",
         "test": "concurrently --success first --kill-others -n eth,backend,test \"yarn start-ganache\" \"yarn start-test-backend\" \"yarn build && mocha --exit dist/js/src/test/\"",
-        "test-ts": "concurrently --success first --kill-others -n eth,backend,test \"yarn start-ganache\" \"yarn start-test-backend\" \"yarn generate-schemas && mocha --exit -r ts-node/register src/test/*.ts\""
+        "test-ts": "concurrently --success first --kill-others -n eth,backend,test \"yarn start-ganache\" \"yarn start-test-backend\" \"yarn generate-schemas && mocha --exit -r ts-node/register src/test/*.ts\"",
+        "clean": "rm -rf build dist"
     },
     "types": "dist/js/src/exports.d.ts",
     "dependencies": {
