+++ packages/market-matcher/tsconfig.build.json
@@ -1,0 +1,8 @@
+{
+    "extends": "../../tsconfig.build.json",
+    "compilerOptions": {
+        "outDir": "./dist/js",
+        "types": ["mocha"]
+    },
+    "include": ["./src/**/*", "./src/test/*"]
+}
