+++ packages/market-matcher/tsconfig.json
@@ -1,8 +1,5 @@
 {
     "extends": "../../tsconfig.json",
     "compilerOptions": {
-        "outDir": "./dist/js",
-        "types": ["mocha"]
-    },
-    "include": ["./src/**/*", "./src/test/*"]
+    }
 }
