+++ packages/market/package.json
@@ -26,7 +26,7 @@
     "build-schema:SupplyPropertiesOffChain": "typescript-json-schema --ignoreErrors --required src/blockchain-facade/Supply.ts ISupplyOffchainProperties > schemas/SupplyOffchainProperties.schema.json",
     "prebuild-schemas": "rm -rf dist/schemas && mkdir -p dist/schemas && rm -rf schemas && mkdir schemas",
     "build-schemas": "npm run build-schema:AgreementPropertiesOffChain && npm run build-schema:MatcherPropertiesOffChain && npm run build-schema:DemandPropertiesOffChain && npm run build-schema:SupplyPropertiesOffChain && cp -R schemas dist/schemas",
-    "build-ts": "npm run build-schemas && rm -rf dist/js && tsc",
+    "build-ts": "npm run build-schemas && rm -rf dist/js && tsc -b tsconfig.build.json",
     "compile": "truffle compile",
     "deploy-contracts": "truffle migrate",
     "flatten-contracts": "truffle-flattener contracts/**/* contracts/MarketContractLookup.sol > build/flattenedContracts.sol && node scripts/removeExperimentalDuplicates.js build/flattenedContracts.sol",
@@ -38,7 +38,8 @@
     "test": "npm run build && mocha dist/js/src/test/ --timeout 60000",
     "test-ts": "mocha -r ts-node/register src/test/*.ts --timeout 60000",
     "prettier": "prettier --write --config-precedence file-override './src/**/*'",
-    "test-contracts": "concurrently --success first --kill-others -n eth,backend,test \"yarn start-ganache\" \"yarn start-test-backend\" \"yarn test\""
+    "test-contracts": "concurrently --success first --kill-others -n eth,backend,test \"yarn start-ganache\" \"yarn start-test-backend\" \"yarn test\"",
+    "clean": "rm -rf build dist"
   },
   "types": "dist/js/src/index.d.ts",
   "dependencies": {
