+++ packages/market/tsconfig.json
@@ -1,7 +1,5 @@
 {
     "extends": "../../tsconfig.json",
     "compilerOptions": {
-        "outDir": "./dist/js"
-    },
-    "include": ["./src/**/**/*"]
+    }
 }
