+++ packages/origin-ui/tsconfig.json
@@ -1,12 +1,5 @@
 {
     "extends": "../../tsconfig.json",
     "compilerOptions": {
-        "outDir": "./dist/",
-        "declaration": false,
-        "moduleResolution": "node",
-        "module": "es2015",
-        "target": "es2018"
-    },
-    "include": ["./src/**/*"],
-    "exclude": ["./src/__tests__/**/*"]
+    }
 }
