+++ packages/origin-ui/webpack/base.config.js
@@ -3,15 +3,15 @@ const HtmlWebpackPlugin = require('html-webpack-plugin');
 const path = require('path');
 
 module.exports = {
-    entry: "./src/index.tsx",
+    entry: './src/index.tsx',
     output: {
-        filename: "bundle.js",
-        path: __dirname + "/../dist"
+        filename: 'bundle.js',
+        path: __dirname + '/../dist'
     },
 
     resolve: {
         // Add '.ts' and '.tsx' as resolvable extensions.
-        extensions: [".ts", ".tsx", ".js", ".json"]
+        extensions: ['.ts', '.tsx', '.js', '.json']
     },
 
     devServer: {
@@ -24,50 +24,58 @@ module.exports = {
         new ExtractTextPlugin({
             filename: 'styles.css',
             allChunks: true
-      }),
-      new HtmlWebpackPlugin({
-        title: 'Origin',
-        favicon: 'favicon.ico',
-        template: './src/index.ejs',
-        meta: {
-            viewport: 'width=device-width, initial-scale=1, shrink-to-fit=no'
-        }
-      })
+        }),
+        new HtmlWebpackPlugin({
+            title: 'Origin',
+            favicon: 'favicon.ico',
+            template: './src/index.ejs',
+            meta: {
+                viewport: 'width=device-width, initial-scale=1, shrink-to-fit=no'
+            }
+        })
     ],
     module: {
-    
         rules: [
             {
                 test: /\.(scss|css)$/,
-                use: [{
-                    loader: "style-loader" // creates style nodes from JS strings
-                }, {
-                    loader: "css-loader" // translates CSS into CommonJS
-                }, {
-                    loader: "sass-loader" // compiles Sass to CSS
-                }]
+                use: [
+                    {
+                        loader: 'style-loader' // creates style nodes from JS strings
+                    },
+                    {
+                        loader: 'css-loader' // translates CSS into CommonJS
+                    },
+                    {
+                        loader: 'sass-loader' // compiles Sass to CSS
+                    }
+                ]
             },
             {
                 test: /\.(png|jpg|gif|svg)$/,
                 use: [
-                  {
-                    loader: 'url-loader',
-                    options: {
-                      limit: 8192
+                    {
+                        loader: 'url-loader',
+                        options: {
+                            limit: 8192
+                        }
                     }
-                  }
                 ]
-              },
+            },
             // All files with a '.ts' or '.tsx' extension will be handled by 'awesome-typescript-loader'.
-            { test: /\.tsx?$/, loader: "ts-loader" },
+            {
+                test: /\.tsx?$/,
+                loader: 'ts-loader',
+                options: {
+                    configFile: '../tsconfig.build.json'
+                }
+            },
 
             // All output '.js' files will have any sourcemaps re-processed by 'source-map-loader'.
-            { enforce: "pre", test: /\.js\.map$/, loader: "source-map-loader" }
+            { enforce: 'pre', test: /\.js\.map$/, loader: 'source-map-loader' }
         ]
-        
     },
 
     node: {
         fs: 'empty'
-      }
+    }
 };
