+++ packages/origin/tsconfig.build.json
@@ -1,0 +1,7 @@
+{
+    "extends": "../../tsconfig.build.json",
+    "compilerOptions": {
+        "outDir": "./dist/js"
+    },
+    "include": ["./src/**/*", "./dist/ts/**/*"]
+}
