+++ packages/solar-simulator/tsconfig.json
@@ -1,12 +1,5 @@
 {
     "extends": "../../tsconfig.json",
     "compilerOptions": {
-        "outDir": "./dist/js",
-        "types": [
-            "node"
-        ]
-    },
-    "include": [
-        "./src/**/*"
-    ]
+    }
 }
