+++ packages/user-registry/tsconfig.build.json
@@ -1,0 +1,11 @@
+{
+    "extends": "../../tsconfig.build.json",
+    "compilerOptions": {
+        "outDir": "./dist/js",
+    },
+    "include": [
+        "./src/**/*",
+        "./dist/ts/**/*",
+        "./build/**/*"
+    ]
+}
