+++ packages/user-registry/tsconfig.json
@@ -1,11 +1,5 @@
 {
     "extends": "../../tsconfig.json",
     "compilerOptions": {
-        "outDir": "./dist/js",
-    },
-    "include": [
-        "./src/**/*",
-        "./dist/ts/**/*",
-        "./build/**/*"
-    ]
+    }
 }
