+++ packages/utils-demo/package.json
@@ -10,7 +10,7 @@
   "scripts": {
     "prebuild-ts": "rm -rf build/ts",
     "build": "npm run build-ts",
-    "build-ts": "rm -rf dist/js && tsc",
+    "build-ts": "rm -rf dist/js && tsc -b tsconfig.build.json",
     "deploy-empty": "npm run build-ts && node dist/js/deployEmpty",
     "lint": "./node_modules/.bin/tslint --fix 'src/**/*{.ts,.tsx}'",
     "start": "npm run build-ts && node dist/js/test",
@@ -18,7 +18,8 @@
     "start-ganache": "ganache-cli -m 'chalk park staff buzz chair purchase wise oak receive avoid avoid home' -l 8000000 -e 1000000 -a 20 -h 0.0.0.0",
     "start-test-backend": "node ../../node_modules/@energyweb/utils-testbackend/dist/js/src/index.js",
     "test": "npm run build-ts && node dist/js/test",
-    "prettier": "prettier --write --config-precedence file-override './src/**/*'"
+    "prettier": "prettier --write --config-precedence file-override './src/**/*'",
+    "clean": "rm -rf dist"
   },
   "types": "dist/js/index.d.ts",
   "dependencies": {
