+++ packages/utils-general/package.json
@@ -22,14 +22,15 @@
   },
   "scripts": {
     "build": "scripts/build.js",
-    "build-ts": "rm -rf dist/js && tsc",
+    "build-ts": "rm -rf dist/js && tsc -b tsconfig.build.json",
     "compile": "truffle compile",
     "deploy-contracts": "truffle migrate",
     "lint": "solium -d contracts && tslint 'src/**/*{.ts,.tsx}'",
     "lint-fix": "solium -d contracts --fix && tslint --fix 'src/**/*{.ts,.tsx}'",
     "prepare": "scripts/build.js",
     "start-ganache": "ganache-cli -m 'chalk park staff buzz chair purchase wise oak receive avoid avoid home' -g 0 -l 80000000 -e 1000000 -a 20",
-    "prettier": "prettier --write --config-precedence file-override './src/**/*'"
+    "prettier": "prettier --write --config-precedence file-override './src/**/*'",
+    "clean": "rm -rf build dist"
   },
   "types": "dist/js/index.d.ts",
   "dependencies": {
