+++ packages/utils-general/tsconfig.json
@@ -1,9 +1,5 @@
 {
     "extends": "../../tsconfig.json",
     "compilerOptions": {
-        "outDir": "./dist/js",
-    },
-    "include": [
-        "src/**/*"
-    ]
+    }
 }
