+++ packages/utils-testbackend/tsconfig.build.json
@@ -1,0 +1,9 @@
+{
+    "extends": "../../tsconfig.build.json",
+    "compilerOptions": {
+        "outDir": "./dist/js",
+    },
+    "include": [
+        "src/**/*"
+    ]
+}
