+++ tsconfig.json
@@ -1,18 +1,9 @@
 {
-  "compilerOptions": {
-      "sourceMap": true,
-      "noImplicitAny": false,
-      "declaration": true,
-      "module": "CommonJS",
-      "target": "es2018",
-      "jsx": "react",
-      "allowSyntheticDefaultImports": true,
-      "resolveJsonModule": true,
-      "esModuleInterop": true,
-      "lib": [
-          "es2018",
-          "dom"
-      ],
-      "skipLibCheck": true
-  }
+    "extends": "./tsconfig.build.json",
+    "compilerOptions": {
+        "baseUrl": "./packages",
+        "paths": {
+            "@energyweb/*": ["*/src"]
+        }
+    }
 }
